Reorganize directory structure for jobs and the framework
